+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 [Full diff](https://github.com/sider/runners/compare/0.46.0...HEAD)
 
+- **Metrics Code Clone** Fix NoMethodError in `Runners::Processor::MetricsCodeClone#initialize` [#2277](https://github.com/sider/runners/pull/2277)
+
 ## 0.46.0
 
 [Full diff](https://github.com/sider/runners/compare/0.45.0...0.46.0)